+++ host/class/uac/usb_host_uac/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog for USB Host UAC
+
+## 1.1.0
+
+### Improvements
+
+- Added add `uac_host_device_open_with_vid_pid` to open connected audio devices with known VID and PID
+- Print component version message `uac-host: Install Succeed, Version: 1.1.0` in `uac_host_install` function
+
 ## 1.0.0
 
 - Initial version